refactor: rename AutoCoralAndElevatorCmd to CoralAutoToReefCommandGroup and update references
