+++ src/main/java/frc/robot/commandGroups/CoralAutoToReefCommandGroup.java
@@ -2,11 +2,13 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commandGroups;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import frc.robot.commands.CoralShooterHoldCmd;
+import frc.robot.commands.TagTrackingCmd;
 import frc.robot.commands.TagTrackingCmd.AimTarget;
 import frc.robot.drivebase.SwerveDrive;
 import frc.robot.subsystems.CoralShooterSubsystem;
@@ -16,12 +18,12 @@ import java.util.Map;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class AutoCoralAndElevatorCmd extends SequentialCommandGroup {
+public class CoralAutoToReefCommandGroup extends SequentialCommandGroup {
   SwerveDrive swerveDrive;
   ElevatorSubsystem elevatorSubsystem;
   CoralShooterSubsystem coralShooterSubsystem;
 
-  public AutoCoralAndElevatorCmd(SwerveDrive swerveDrive, ElevatorSubsystem elevatorSubsystem,
+  public CoralAutoToReefCommandGroup(SwerveDrive swerveDrive, ElevatorSubsystem elevatorSubsystem,
       CoralShooterSubsystem coralShooterSubsystem, int targetFloor, Boolean isLeft, Boolean isAutoTime) {
     this.swerveDrive = swerveDrive;
     this.elevatorSubsystem = elevatorSubsystem;
